6.7.0 - Alpha12 - Fine tuning

+++ app/src/main/java/org/autojs/autojs/network/entity/SimpleVersionInfo.kt
@@ -1,3 +1,3 @@
 package org.autojs.autojs.network.entity
 
-class SimpleVersionInfo(val versionName: String?, val versionCode: Int)
+class SimpleVersionInfo(val versionName: String?, val versionCode: Long)
